Move cosmos token adapter out
