+++ typescript/sdk/src/token/adapters/CosmWasmTokenAdapter.ts
@@ -1,6 +1,5 @@
 import { ExecuteInstruction } from '@cosmjs/cosmwasm-stargate';
-import { Coin, MsgTransferEncodeObject } from '@cosmjs/stargate';
-import { MsgTransfer } from 'cosmjs-types/ibc/applications/transfer/v1/tx';
+import { Coin } from '@cosmjs/stargate';
 
 import {
   Address,
@@ -9,10 +8,7 @@ import {
   strip0x,
 } from '@hyperlane-xyz/utils';
 
-import {
-  BaseCosmWasmAdapter,
-  BaseCosmosAdapter,
-} from '../../app/MultiProtocolApp';
+import { BaseCosmWasmAdapter } from '../../app/MultiProtocolApp';
 import {
   BalanceResponse,
   ExecuteMsg as Cw20Execute,
@@ -32,7 +28,7 @@ import {
 } from '../../cw-types/WarpCw20.types';
 import { MultiProtocolProvider } from '../../providers/MultiProtocolProvider';
 import { ChainName } from '../../types';
-import { ERC20Metadata, MinimalTokenMetadata } from '../config';
+import { ERC20Metadata } from '../config';
 
 import {
   IHypTokenAdapter,
@@ -40,61 +36,6 @@ import {
   TransferParams,
   TransferRemoteParams,
 } from './ITokenAdapter';
-
-// Interacts with IBC denom tokens
-export class NativeTokenAdapter
-  extends BaseCosmosAdapter
-  implements ITokenAdapter
-{
-  constructor(
-    public readonly chainName: string,
-    public readonly multiProvider: MultiProtocolProvider,
-    public readonly addresses: Record<string, Address>,
-    public readonly ibcDenom: string = 'untrn',
-  ) {
-    super(chainName, multiProvider, addresses);
-  }
-
-  async getBalance(address: string): Promise<string> {
-    const provider = await this.getProvider();
-    const coin = await provider.getBalance(address, this.ibcDenom);
-    return coin.amount;
-  }
-
-  getMetadata(): Promise<MinimalTokenMetadata> {
-    throw new Error('Metadata not available to native tokens');
-  }
-
-  populateApproveTx(_transferParams: TransferParams): unknown {
-    throw new Error('Approve not required for native tokens');
-  }
-
-  async populateTransferTx(
-    transferParams: TransferParams,
-  ): Promise<MsgTransferEncodeObject> {
-    const transfer: MsgTransfer = {
-      sourcePort: '',
-      sourceChannel: '',
-      token: {
-        denom: this.ibcDenom,
-        amount: transferParams.weiAmountOrId.toString(),
-      },
-      sender: '',
-      receiver: '',
-      timeoutHeight: {
-        revisionNumber: 0n,
-        revisionHeight: 0n,
-      },
-      timeoutTimestamp: 0n,
-      memo: '', // how to encode this?
-    };
-    return {
-      typeUrl: '/ibc.applications.transfer.v1.MsgTransfer',
-      // @ts-ignore
-      value: transfer,
-    };
-  }
-}
 
 // Interacts with IBC denom tokens in CosmWasm
 export class CwNativeTokenAdapter
